+++ core/src/main/java/me/hsgamer/topper/core/TopSnapshot.java
@@ -1,13 +1,12 @@
 package me.hsgamer.topper.core;
 
-import java.math.BigDecimal;
 import java.util.UUID;
 
 public class TopSnapshot {
     public final UUID uuid;
-    public final BigDecimal value;
+    public final double value;
 
-    TopSnapshot(UUID uuid, BigDecimal value) {
+    TopSnapshot(UUID uuid, double value) {
         this.uuid = uuid;
         this.value = value;
     }
@@ -16,7 +15,7 @@ public class TopSnapshot {
         return uuid;
     }
 
-    public BigDecimal getValue() {
+    public double getValue() {
         return value;
     }
 }
