+++ core/src/main/java/me/hsgamer/topper/core/TopStorage.java
@@ -1,6 +1,5 @@
 package me.hsgamer.topper.core;
 
-import java.math.BigDecimal;
 import java.util.Map;
 import java.util.UUID;
 import java.util.concurrent.CompletableFuture;
@@ -9,7 +8,7 @@ import java.util.logging.Logger;
 public interface TopStorage {
     Logger LOGGER = Logger.getLogger(TopStorage.class.getName());
 
-    CompletableFuture<Map<UUID, BigDecimal>> load(TopHolder holder);
+    CompletableFuture<Map<UUID, Double>> load(TopHolder holder);
 
     CompletableFuture<Void> save(TopEntry topEntry, boolean onUnregister);
